==> Data Cleaning and Fixing/11v11 Data/step0-analyse-all-match-data.ipynb <==
import pandas as pd
import numpy as np
import time

df = pd.read_csv("all_match_data.csv")
len(df)

refs = df['Referee'].tolist()
unique_refs = set(refs)
len(unique_refs)

unique_refs

def check_if_ref_name(row,ref_name):
    ref = row['Referee']
    if ref == ref_name:
        return True
    else:
        return False

df['RefisPrem'] = df.apply(check_if_ref_name,args=('Premier League ',),axis=1)

df_ref_prem = df.query('RefisPrem == True')

len(df_ref_prem)

df_ref_prem['Data Link'].tolist()

# looking at these, all dates are pre 2000, and the dates seems to be in batches of 4 (4 games each week). 
# This, alongside checking a few random links and googling elsewhere for who the ref actually was, 
# imply to me that there may potentially be 4 referee's here that for some reason were never mentioned of.

# We'll investigate this by looking at the last (most recent) 12 games, and researching their ref's:
#'/matches/newcastle-united-v-middlesbrough-01-may-1999-24056/', - Mike Reed
#'/matches/nottingham-forest-v-sheffield-wednesday-01-may-1999-24057/', - Jeff Winter
#'/matches/southampton-v-leicester-city-01-may-1999-24058/', - Dermot Gallagher
#'/matches/west-ham-united-v-leeds-united-01-may-1999-24059/', - Rob Harris
#'/matches/liverpool-v-manchester-united-05-may-1999-24063/', - David Elleray
#'/matches/leicester-city-v-newcastle-united-08-may-1999-24069/', - Uriah Rennie
#'/matches/sheffield-wednesday-v-liverpool-08-may-1999-24070/', - Peter Jones
#'/matches/middlesbrough-v-manchester-united-09-may-1999-24072/', - Graham Barber
#'/matches/coventry-city-v-leeds-united-16-may-1999-24079/', - Stephen Lodge
#'/matches/manchester-united-v-tottenham-hotspur-16-may-1999-24081/', - Graham Poll
#'/matches/nottingham-forest-v-leicester-city-16-may-1999-24083/', - Steve Dunn
#'/matches/southampton-v-everton-16-may-1999-24084/' - Graham Barber
