==> Data Cleaning and Fixing/11v11 Data/step1-fix-referee-data.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("all_match_data_raw_copy.csv")
%run list_of_games_without_referees.ipynb

elevendataRefs = df['Referee'].tolist()
elevendataDataLinks = df['Data Link'].tolist()
elevendataTeams = [x.split("/")[2].split("-")[:-4] for x in elevendataDataLinks]
elevendataTeams[0:5]

df_fix_data = pd.read_csv("referee_fix_data.csv")

premdataRefs = df_fix_data['Referee'].tolist()
premdataHomeTeams = df_fix_data['HomeTeam'].tolist()
premdataAwayTeams = df_fix_data['AwayTeam'].tolist()

def select_home_team(match):
    homeTeam = match[0]
    if match[1] == 'v':
        return homeTeam
    elif match[2] == 'v':
        return homeTeam + " " + match[1]
    elif match[3] == 'v':
        return homeTeam + " " + match[1] + " " + match[2]
    elif match[4] == 'v':
        return homeTeam + " " + match[1] + " " + match[2] + " " + match[3]
    else:
        print("Error? : {}".format(match))
        
def select_away_team(match):
    awayTeam = match[-1]
    if match[-2] == 'v':
        return awayTeam
    elif match[-3] == 'v':
        return match[-2] + " " + awayTeam
    elif match[-4] == 'v':
        return match[-3] + " " + match[-2] + " " + awayTeam
    elif match[-5] == 'v':
        return match[-4] + " " + match[-3] + " " + match[-2] + " " + awayTeam
    else:    
        print("Error? : {}".format(match))

elevendataHomeTeams = [select_home_team(x) for x in elevendataTeams]
elevendataAwayTeams = [select_away_team(x) for x in elevendataTeams]

def fix_team_names(x):
    if "queens" in x:
        return "qpr"
    else:
        return x


elevendataHomeTeams = [fix_team_names(x) for x in elevendataHomeTeams]
elevendataAwayTeams = [fix_team_names(x) for x in elevendataAwayTeams]

j = 0
skipFirst = 6
max_j = len(premdataRefs)-1
for i in range(len(elevendataRefs)):
    eRef = elevendataRefs[i]
    eHome = elevendataHomeTeams[i]
    eAway = elevendataAwayTeams[i]
    if eRef != 'Premier League ':
        continue
    elif skipFirst > 0:
        skipFirst -= 1
        continue
    print("Fixing: {}: {} v {}".format(i,eHome,eAway))
    notfound = True
    while notfound and j <= max_j:
        pHome = premdataHomeTeams[j]
        pAway = premdataAwayTeams[j]
        print("Comparing to {} v {}".format(pHome,pAway))
        if eHome.lower() in pHome.lower() and eAway.lower() in pAway.lower():
            elevendataRefs[i] = premdataRefs[j]
            notfound = False
        j += 1
    print()

len(elevendataRefs)

set(elevendataRefs)

# Steve lodge = stephen lodge?
# mike riley = michael riley?
# keith cooper = keith A cooper?
df['Referee'] = elevendataRefs

df_gagelmann = df.query("Referee ==  ' Gagelmann'")
df_gagelmann['Data Link'].tolist()

# this match turns out to have Dermot Gallagher as ref instead, so will replace
def remove_gagelmann(row):
    ref = row['Referee']
    if ref == ' Gagelmann':
        print("Found it!")
        return "Dermot Gallagher"
    else:
        return ref
df['Referee'] = df.apply(remove_gagelmann,axis=1)

# Now need to investigate the missing refs
df_steve_lodge = df.query("Referee == 'Steve Lodge'")
df_stephen_lodge = df.query("Referee == 'Stephen Lodge'")
print("{} v {}".format(len(df_steve_lodge),len(df_stephen_lodge)))

df_mike_riley = df.query("Referee == 'Mike Riley'")
df_michael_riley = df.query("Referee == 'Michael Riley'")
print("{} v {}".format(len(df_mike_riley),len(df_michael_riley)))

df_keith_cooper = df.query("Referee == 'Keith Cooper'")
df_keith_a_cooper = df.query("Referee == 'Keith A Cooper'")
print("{} v {}".format(len(df_keith_cooper),len(df_keith_a_cooper)))

df_keith_a_cooper.iloc[18].tolist()[0]

def replace_ref_names(row):
    ref = row['Referee']
    if ref == 'Stephen Lodge':
        return "Steve Lodge"
    elif ref == 'Michael Riley':
        return "Mike Riley"
    elif ref == 'Keith A Cooper':
        return "Keith Cooper"
    else:
        return ref
df['Referee'] = df.apply(replace_ref_names,axis=1)

from collections import defaultdict
ref_dict = defaultdict(int)
for ref in df['Referee'].tolist():
    ref_dict[ref] += 1
rare_refs = defaultdict(int)
for ref in ref_dict.keys():
    count = ref_dict[ref]
    if count < 10:
        rare_refs[ref] = count
rare_refs

df_specific_ref = df.query("Referee == 'David Coote'")
df_specific_ref['Data Link'].tolist()

df.to_csv("all_match_data_fix1.csv",index=False)
