==> Data Cleaning and Fixing/11v11 Data/step1-get-data-for-null-referees.ipynb <==
import pandas as pd
import numpy as np
import time
import requests
from bs4 import BeautifulSoup as soup

response = requests.get("https://www.premierleague.com/match/1393",headers={'User-Agent': 'Custom'})
page = soup(response.content, "html.parser")

def get_data_for_link(url,number):
    response = requests.get(url+str(number),headers={'User-Agent': 'Custom'})
    page = soup(response.content, "html.parser")
    referees = page.findAll('div',{'class':'referee'})
    ref = referees[0].text.strip()
    homeTeams = page.findAll('div',{'class':'team home'})
    homeTeam = homeTeams[0].findAll('span',{'class':'long'})[0].text
    awayTeams = page.findAll('div',{'class':'team away'})
    awayTeam = awayTeams[0].findAll('span',{'class':'long'})[0].text
    return (number,homeTeam,awayTeam,ref)

(num,homeTeam,awayTeam,ref) = get_data_for_link("https://www.premierleague.com/match/",1393)
print("{}: {} v {} - Ref: {}".format(num,homeTeam,awayTeam,ref))

def get_ref_dataset(base_url,start_num,end_num):
    rows_list = []
    for i in range(start_num,end_num+1):
        data_tuple = get_data_for_link(base_url,i)
        dict1 = {}
        dict1['ID'] = data_tuple[0]
        dict1['HomeTeam'] = data_tuple[1]
        dict1['AwayTeam'] = data_tuple[2]
        dict1['Referee'] = data_tuple[3]
        rows_list.append(dict1)
        print(i)
    df_all_matches = pd.DataFrame(data=rows_list,columns=['ID','HomeTeam','AwayTeam','Referee'])
    df_all_matches.to_csv("referee_fix_data.csv",index=False)

get_ref_dataset("https://www.premierleague.com/match/",2153,2905)
