==> Data Cleaning and Fixing/11v11 Data/step1-list-of-games-without-referees.ipynb <==
url_list = ['/matches/southampton-v-nottingham-forest-19-august-1995-22179/',
 '/matches/wimbledon-v-queens-park-rangers-20-january-1996-22404/',
 '/matches/blackburn-rovers-v-leeds-united-04-september-1996-22585/',
 '/matches/chelsea-v-tottenham-hotspur-26-october-1996-22653/',
 '/matches/aston-villa-v-nottingham-forest-02-november-1996-22662/',
 '/matches/blackburn-rovers-v-sunderland-01-march-1997-22814/',
 '/matches/newcastle-united-v-sheffield-wednesday-09-august-1997-22939/',
 '/matches/liverpool-v-leicester-city-13-august-1997-22947/',
 '/matches/manchester-united-v-southampton-13-august-1997-22948/',
 '/matches/sheffield-wednesday-v-leeds-united-13-august-1997-22949/',
 '/matches/west-ham-united-v-tottenham-hotspur-13-august-1997-22950/',
 '/matches/everton-v-west-ham-united-23-august-1997-22953/',
 '/matches/leicester-city-v-manchester-united-23-august-1997-22955/',
 '/matches/tottenham-hotspur-v-derby-county-23-august-1997-22958/',
 '/matches/wimbledon-v-sheffield-wednesday-23-august-1997-22959/',
 '/matches/leeds-united-v-liverpool-26-august-1997-22963/',
 '/matches/everton-v-manchester-united-27-august-1997-22965/',
 '/matches/sheffield-wednesday-v-leicester-city-30-august-1997-22976/',
 '/matches/west-ham-united-v-wimbledon-30-august-1997-22977/',
 '/matches/leicester-city-v-tottenham-hotspur-13-september-1997-22985/',
 '/matches/liverpool-v-sheffield-wednesday-13-september-1997-22986/',
 '/matches/manchester-united-v-west-ham-united-13-september-1997-22987/',
 '/matches/newcastle-united-v-wimbledon-13-september-1997-22988/',
 '/matches/leeds-united-v-leicester-city-20-september-1997-22993/',
 '/matches/southampton-v-liverpool-20-september-1997-22995/',
 '/matches/west-ham-united-v-newcastle-united-20-september-1997-22997/',
 '/matches/southampton-v-leeds-united-24-september-1997-23009/',
 '/matches/derby-county-v-southampton-27-september-1997-23014/',
 '/matches/leeds-united-v-manchester-united-27-september-1997-23016/',
 '/matches/west-ham-united-v-liverpool-27-september-1997-23018/',
 '/matches/newcastle-united-v-tottenham-hotspur-04-october-1997-23024/',
 '/matches/southampton-v-west-ham-united-04-october-1997-23026/',
 '/matches/leicester-city-v-derby-county-06-october-1997-23029/',
 '/matches/everton-v-liverpool-18-october-1997-23035/',
 '/matches/leeds-united-v-newcastle-united-18-october-1997-23036/',
 '/matches/tottenham-hotspur-v-sheffield-wednesday-19-october-1997-23038/',
 '/matches/liverpool-v-derby-county-25-october-1997-23041/',
 '/matches/southampton-v-tottenham-hotspur-25-october-1997-23045/',
 '/matches/wimbledon-v-leeds-united-25-october-1997-23046/',
 '/matches/leicester-city-v-west-ham-united-27-october-1997-23049/',
 '/matches/barnsley-v-blackburn-rovers-01-november-1997-23051/',
 '/matches/manchester-united-v-sheffield-wednesday-01-november-1997-23054/',
 '/matches/newcastle-united-v-leicester-city-01-november-1997-23055/',
 '/matches/tottenham-hotspur-v-leeds-united-01-november-1997-23056/',
 '/matches/everton-v-southampton-02-november-1997-23058/',
 '/matches/leeds-united-v-derby-county-08-november-1997-23063/',
 '/matches/liverpool-v-tottenham-hotspur-08-november-1997-23064/',
 '/matches/leicester-city-v-wimbledon-10-november-1997-23069/',
 '/matches/newcastle-united-v-southampton-22-november-1997-23075/',
 '/matches/leeds-united-v-west-ham-united-23-november-1997-23078/',
 '/matches/everton-v-tottenham-hotspur-29-november-1997-23086/',
 '/matches/southampton-v-sheffield-wednesday-29-november-1997-23088/',
 '/matches/derby-county-v-west-ham-united-06-december-1997-23094/',
 '/matches/leeds-united-v-everton-06-december-1997-23095/',
 '/matches/liverpool-v-manchester-united-06-december-1997-23097/',
 '/matches/wimbledon-v-southampton-07-december-1997-23100/',
 '/matches/everton-v-wimbledon-13-december-1997-23107/',
 '/matches/southampton-v-leicester-city-13-december-1997-23108/',
 '/matches/west-ham-united-v-sheffield-wednesday-13-december-1997-23109/',
 '/matches/newcastle-united-v-derby-county-17-december-1997-23112/',
 '/matches/leicester-city-v-everton-20-december-1997-23117/',
 '/matches/newcastle-united-v-manchester-united-21-december-1997-23121/',
 '/matches/derby-county-v-newcastle-united-26-december-1997-23128/',
 '/matches/liverpool-v-leeds-united-26-december-1997-23129/',
 '/matches/manchester-united-v-everton-26-december-1997-23130/',
 '/matches/leicester-city-v-sheffield-wednesday-28-december-1997-23138/',
 '/matches/newcastle-united-v-liverpool-28-december-1997-23139/',
 '/matches/wimbledon-v-west-ham-united-28-december-1997-23141/',
 '/matches/arsenal-v-leeds-united-10-january-1998-23143/',
 '/matches/crystal-palace-v-everton-10-january-1998-23147/',
 '/matches/liverpool-v-wimbledon-10-january-1998-23149/',
 '/matches/manchester-united-v-tottenham-hotspur-10-january-1998-23150/',
 '/matches/sheffield-wednesday-v-newcastle-united-10-january-1998-23151/',
 '/matches/leeds-united-v-sheffield-wednesday-17-january-1998-23157/',
 '/matches/leicester-city-v-liverpool-17-january-1998-23158/',
 '/matches/tottenham-hotspur-v-west-ham-united-17-january-1998-23161/',
 '/matches/wimbledon-v-derby-county-17-january-1998-23162/',
 '/matches/southampton-v-manchester-united-19-january-1998-23160/',
 '/matches/liverpool-v-newcastle-united-20-january-1998-22978/',
 '/matches/crystal-palace-v-leeds-united-31-january-1998-23167/',
 '/matches/derby-county-v-tottenham-hotspur-31-january-1998-23168/',
 '/matches/manchester-united-v-leicester-city-31-january-1998-23170/',
 '/matches/sheffield-wednesday-v-wimbledon-31-january-1998-23171/',
 '/matches/west-ham-united-v-everton-31-january-1998-23172/',
 '/matches/leicester-city-v-leeds-united-07-february-1998-23179/',
 '/matches/liverpool-v-southampton-07-february-1998-23180/',
 '/matches/newcastle-united-v-west-ham-united-07-february-1998-23182/',
 '/matches/arsenal-v-chelsea-08-february-1998-23173/',
 '/matches/crystal-palace-v-wimbledon-09-february-1998-23177/',
 '/matches/everton-v-derby-county-14-february-1998-23186/',
 '/matches/sheffield-wednesday-v-liverpool-14-february-1998-23188/',
 '/matches/tottenham-hotspur-v-leicester-city-14-february-1998-23190/',
 '/matches/manchester-united-v-derby-county-21-february-1998-23198/',
 '/matches/newcastle-united-v-leeds-united-22-february-1998-23199/',
 '/matches/everton-v-newcastle-united-28-february-1998-23209/',
 '/matches/leeds-united-v-southampton-28-february-1998-23210/',
 '/matches/leeds-united-v-tottenham-hotspur-04-march-1998-23218/',
 '/matches/sheffield-wednesday-v-manchester-united-07-march-1998-23221/',
 '/matches/tottenham-hotspur-v-liverpool-14-march-1998-23230/',
 '/matches/wimbledon-v-leicester-city-14-march-1998-23232/',
 '/matches/crystal-palace-v-tottenham-hotspur-28-march-1998-23238/',
 '/matches/manchester-united-v-wimbledon-28-march-1998-23240/',
 '/matches/southampton-v-newcastle-united-28-march-1998-23241/',
 '/matches/west-ham-united-v-leeds-united-30-march-1998-23242/',
 '/matches/wimbledon-v-newcastle-united-31-march-1998-23192/',
 '/matches/sheffield-wednesday-v-southampton-04-april-1998-23250/',
 '/matches/tottenham-hotspur-v-everton-04-april-1998-23251/',
 '/matches/manchester-united-v-liverpool-10-april-1998-23260/',
 '/matches/crystal-palace-v-leicester-city-11-april-1998-23258/',
 '/matches/everton-v-leeds-united-11-april-1998-23259/',
 '/matches/southampton-v-wimbledon-11-april-1998-23261/',
 '/matches/west-ham-united-v-derby-county-11-april-1998-23262/',
 '/matches/sheffield-wednesday-v-west-ham-united-13-april-1998-23270/',
 '/matches/everton-v-leicester-city-18-april-1998-23279/',
 '/matches/manchester-united-v-newcastle-united-18-april-1998-23280/',
 '/matches/chelsea-v-sheffield-wednesday-19-april-1998-23276/',
 '/matches/everton-v-sheffield-wednesday-25-april-1998-23289/',
 '/matches/tottenham-hotspur-v-newcastle-united-25-april-1998-23291/',
 '/matches/derby-county-v-leicester-city-26-april-1998-23288/',
 '/matches/crystal-palace-v-manchester-united-27-april-1998-23287/',
 '/matches/leicester-city-v-newcastle-united-29-april-1998-23219/',
 '/matches/liverpool-v-west-ham-united-02-may-1998-23297/',
 '/matches/wimbledon-v-tottenham-hotspur-02-may-1998-23302/',
 '/matches/manchester-united-v-leeds-united-04-may-1998-23298/',
 '/matches/crystal-palace-v-west-ham-united-05-may-1998-23217/',
 '/matches/derby-county-v-liverpool-10-may-1998-23308/',
 '/matches/leeds-united-v-wimbledon-10-may-1998-23310/',
 '/matches/tottenham-hotspur-v-southampton-10-may-1998-23311/',
 '/matches/west-ham-united-v-leicester-city-10-may-1998-23312/',
 '/matches/middlesbrough-v-leeds-united-15-august-1998-23712/',
 '/matches/sheffield-wednesday-v-west-ham-united-15-august-1998-23714/',
 '/matches/southampton-v-liverpool-16-august-1998-23716/',
 '/matches/derby-county-v-wimbledon-22-august-1998-23720/',
 '/matches/leicester-city-v-everton-22-august-1998-23721/',
 '/matches/nottingham-forest-v-coventry-city-22-august-1998-23723/',
 '/matches/tottenham-hotspur-v-sheffield-wednesday-22-august-1998-23724/',
 '/matches/west-ham-united-v-manchester-united-22-august-1998-23725/',
 '/matches/coventry-city-v-west-ham-united-29-august-1998-23730/',
 '/matches/everton-v-tottenham-hotspur-29-august-1998-23731/',
 '/matches/middlesbrough-v-derby-county-29-august-1998-23732/',
 '/matches/southampton-v-nottingham-forest-29-august-1998-23734/',
 '/matches/wimbledon-v-leeds-united-29-august-1998-23735/',
 '/matches/newcastle-united-v-liverpool-30-august-1998-23736/',
 '/matches/leeds-united-v-southampton-08-september-1998-23742/',
 '/matches/nottingham-forest-v-everton-08-september-1998-23738/',
 '/matches/derby-county-v-sheffield-wednesday-09-september-1998-23741/',
 '/matches/leicester-city-v-middlesbrough-09-september-1998-23743/',
 '/matches/liverpool-v-coventry-city-09-september-1998-23744/',
 '/matches/everton-v-leeds-united-12-september-1998-23750/',
 '/matches/manchester-united-v-coventry-city-12-september-1998-23752/',
 '/matches/newcastle-united-v-southampton-12-september-1998-23753/',
 '/matches/west-ham-united-v-liverpool-12-september-1998-23755/',
 '/matches/tottenham-hotspur-v-middlesbrough-13-september-1998-23756/',
 '/matches/coventry-city-v-newcastle-united-19-september-1998-23757/',
 '/matches/derby-county-v-leicester-city-19-september-1998-23758/',
 '/matches/middlesbrough-v-everton-19-september-1998-23761/',
 '/matches/nottingham-forest-v-west-ham-united-19-september-1998-23762/',
 '/matches/southampton-v-tottenham-hotspur-19-september-1998-23763/',
 '/matches/wimbledon-v-sheffield-wednesday-19-september-1998-23764/',
 '/matches/manchester-united-v-liverpool-24-september-1998-23767/',
 '/matches/newcastle-united-v-nottingham-forest-26-september-1998-23772/',
 '/matches/tottenham-hotspur-v-leeds-united-26-september-1998-23774/',
 '/matches/leicester-city-v-wimbledon-27-september-1998-23775/',
 '/matches/west-ham-united-v-southampton-28-september-1998-23776/',
 '/matches/derby-county-v-tottenham-hotspur-03-october-1998-23779/',
 '/matches/leeds-united-v-leicester-city-03-october-1998-23780/',
 '/matches/middlesbrough-v-sheffield-wednesday-03-october-1998-23781/',
 '/matches/southampton-v-manchester-united-03-october-1998-23783/',
 '/matches/wimbledon-v-everton-03-october-1998-23784/',
 '/matches/everton-v-liverpool-17-october-1998-23789/',
 '/matches/manchester-united-v-wimbledon-17-october-1998-23790/',
 '/matches/newcastle-united-v-derby-county-17-october-1998-23792/',
 '/matches/nottingham-forest-v-leeds-united-17-october-1998-23793/',
 '/matches/coventry-city-v-sheffield-wednesday-18-october-1998-23795/',
 '/matches/leicester-city-v-tottenham-hotspur-19-october-1998-23796/',
 '/matches/derby-county-v-manchester-united-24-october-1998-23799/',
 '/matches/liverpool-v-nottingham-forest-24-october-1998-23800/',
 '/matches/sheffield-wednesday-v-everton-24-october-1998-23801/',
 '/matches/southampton-v-coventry-city-24-october-1998-23802/',
 '/matches/tottenham-hotspur-v-newcastle-united-24-october-1998-23803/',
 '/matches/wimbledon-v-middlesbrough-24-october-1998-23804/',
 '/matches/derby-county-v-leeds-united-31-october-1998-23808/',
 '/matches/everton-v-manchester-united-31-october-1998-23809/',
 '/matches/leicester-city-v-liverpool-31-october-1998-23810/',
 '/matches/newcastle-united-v-west-ham-united-31-october-1998-23811/',
 '/matches/sheffield-wednesday-v-southampton-31-october-1998-23812/',
 '/matches/middlesbrough-v-nottingham-forest-01-november-1998-23814/',
 '/matches/liverpool-v-derby-county-07-november-1998-23819/',
 '/matches/nottingham-forest-v-wimbledon-07-november-1998-23820/',
 '/matches/southampton-v-middlesbrough-07-november-1998-23821/',
 '/matches/leeds-united-v-sheffield-wednesday-08-november-1998-23823/',
 '/matches/manchester-united-v-newcastle-united-08-november-1998-23824/',
 '/matches/liverpool-v-leeds-united-14-november-1998-23829/',
 '/matches/newcastle-united-v-sheffield-wednesday-14-november-1998-23831/',
 '/matches/west-ham-united-v-leicester-city-14-november-1998-23833/',
 '/matches/coventry-city-v-everton-15-november-1998-23834/',
 '/matches/nottingham-forest-v-derby-county-16-november-1998-23835/',
 '/matches/sheffield-wednesday-v-manchester-united-21-november-1998-23841/',
 '/matches/tottenham-hotspur-v-nottingham-forest-21-november-1998-23842/',
 '/matches/derby-county-v-west-ham-united-22-november-1998-23844/',
 '/matches/everton-v-newcastle-united-23-november-1998-23845/',
 '/matches/coventry-city-v-leicester-city-28-november-1998-23848/',
 '/matches/newcastle-united-v-wimbledon-28-november-1998-23849/',
 '/matches/west-ham-united-v-tottenham-hotspur-28-november-1998-23852/',
 '/matches/manchester-united-v-leeds-united-29-november-1998-23855/',
 '/matches/leicester-city-v-southampton-05-december-1998-23861/',
 '/matches/tottenham-hotspur-v-liverpool-05-december-1998-23862/',
 '/matches/wimbledon-v-coventry-city-05-december-1998-23863/',
 '/matches/middlesbrough-v-newcastle-united-06-december-1998-23864/',
 '/matches/sheffield-wednesday-v-nottingham-forest-07-december-1998-23865/',
 '/matches/everton-v-southampton-12-december-1998-23869/',
 '/matches/leicester-city-v-nottingham-forest-12-december-1998-23870/',
 '/matches/middlesbrough-v-west-ham-united-12-december-1998-23871/',
 '/matches/tottenham-hotspur-v-manchester-united-12-december-1998-23873/',
 '/matches/wimbledon-v-liverpool-13-december-1998-23875/',
 '/matches/leeds-united-v-coventry-city-14-december-1998-23876/',
 '/matches/coventry-city-v-derby-county-19-december-1998-23879/',
 '/matches/liverpool-v-sheffield-wednesday-19-december-1998-23880/',
 '/matches/manchester-united-v-middlesbrough-19-december-1998-23881/',
 '/matches/southampton-v-wimbledon-19-december-1998-23884/',
 '/matches/west-ham-united-v-everton-19-december-1998-23885/',
 '/matches/coventry-city-v-tottenham-hotspur-26-december-1998-23890/',
 '/matches/everton-v-derby-county-26-december-1998-23891/',
 '/matches/manchester-united-v-nottingham-forest-26-december-1998-23892/',
 '/matches/middlesbrough-v-liverpool-26-december-1998-23893/',
 '/matches/newcastle-united-v-leeds-united-26-december-1998-23894/',
 '/matches/sheffield-wednesday-v-leicester-city-26-december-1998-23895/',
 '/matches/derby-county-v-middlesbrough-28-december-1998-23900/',
 '/matches/liverpool-v-newcastle-united-28-december-1998-23902/',
 '/matches/nottingham-forest-v-southampton-28-december-1998-23903/',
 '/matches/tottenham-hotspur-v-everton-28-december-1998-23904/',
 '/matches/west-ham-united-v-coventry-city-28-december-1998-23905/',
 '/matches/leeds-united-v-wimbledon-29-december-1998-23907/',
 '/matches/coventry-city-v-nottingham-forest-09-january-1999-23910/',
 '/matches/everton-v-leicester-city-09-january-1999-23911/',
 '/matches/wimbledon-v-derby-county-09-january-1999-23916/',
 '/matches/manchester-united-v-west-ham-united-10-january-1999-23917/',
 '/matches/leeds-united-v-middlesbrough-16-january-1999-23920/',
 '/matches/leicester-city-v-manchester-united-16-january-1999-23921/',
 '/matches/liverpool-v-southampton-16-january-1999-23922/',
 '/matches/tottenham-hotspur-v-wimbledon-16-january-1999-23924/',
 '/matches/west-ham-united-v-sheffield-wednesday-16-january-1999-23925/',
 '/matches/coventry-city-v-liverpool-30-january-1999-23929/',
 '/matches/everton-v-nottingham-forest-30-january-1999-23930/',
 '/matches/sheffield-wednesday-v-derby-county-30-january-1999-23933/',
 '/matches/southampton-v-leeds-united-30-january-1999-23934/',
 '/matches/wimbledon-v-west-ham-united-30-january-1999-23935/',
 '/matches/manchester-united-v-derby-county-03-february-1999-23938/',
 '/matches/leeds-united-v-newcastle-united-06-february-1999-23941/',
 '/matches/leicester-city-v-sheffield-wednesday-06-february-1999-23942/',
 '/matches/liverpool-v-middlesbrough-06-february-1999-23943/',
 '/matches/nottingham-forest-v-manchester-united-06-february-1999-23944/',
 '/matches/tottenham-hotspur-v-coventry-city-06-february-1999-23945/',
 '/matches/derby-county-v-everton-07-february-1999-23947/',
 '/matches/west-ham-united-v-nottingham-forest-13-february-1999-23950/',
 '/matches/everton-v-middlesbrough-17-february-1999-23953/',
 '/matches/newcastle-united-v-coventry-city-17-february-1999-23955/',
 '/matches/coventry-city-v-manchester-united-20-february-1999-23958/',
 '/matches/leeds-united-v-everton-20-february-1999-23960/',
 '/matches/liverpool-v-west-ham-united-20-february-1999-23961/',
 '/matches/middlesbrough-v-tottenham-hotspur-20-february-1999-23962/',
 '/matches/everton-v-wimbledon-27-february-1999-23969/',
 '/matches/sheffield-wednesday-v-middlesbrough-27-february-1999-23971/',
 '/matches/tottenham-hotspur-v-derby-county-27-february-1999-23972/',
 '/matches/leicester-city-v-leeds-united-01-march-1999-23975/',
 '/matches/tottenham-hotspur-v-southampton-02-march-1999-24735/',
 '/matches/sheffield-wednesday-v-wimbledon-03-march-1999-23976/',
 '/matches/southampton-v-west-ham-united-06-march-1999-23978/',
 '/matches/wimbledon-v-leicester-city-06-march-1999-23979/',
 '/matches/leeds-united-v-tottenham-hotspur-10-march-1999-23983/',
 '/matches/nottingham-forest-v-newcastle-united-10-march-1999-23984/',
 '/matches/derby-county-v-liverpool-13-march-1999-23987/',
 '/matches/newcastle-united-v-manchester-united-13-march-1999-23990/',
 '/matches/sheffield-wednesday-v-leeds-united-13-march-1999-23991/',
 '/matches/wimbledon-v-nottingham-forest-13-march-1999-23993/',
 '/matches/middlesbrough-v-southampton-14-march-1999-23994/',
 '/matches/southampton-v-sheffield-wednesday-20-march-1999-23999/',
 '/matches/west-ham-united-v-newcastle-united-20-march-1999-24000/',
 '/matches/manchester-united-v-everton-21-march-1999-24002/',
 '/matches/derby-county-v-newcastle-united-03-april-1999-24006/',
 '/matches/leeds-united-v-nottingham-forest-03-april-1999-24007/',
 '/matches/liverpool-v-everton-03-april-1999-24008/',
 '/matches/tottenham-hotspur-v-leicester-city-03-april-1999-24011/',
 '/matches/coventry-city-v-southampton-05-april-1999-24013/',
 '/matches/everton-v-sheffield-wednesday-05-april-1999-24014/',
 '/matches/middlesbrough-v-wimbledon-05-april-1999-24015/',
 '/matches/newcastle-united-v-tottenham-hotspur-05-april-1999-24016/',
 '/matches/nottingham-forest-v-liverpool-05-april-1999-24017/',
 '/matches/derby-county-v-nottingham-forest-10-april-1999-24022/',
 '/matches/leicester-city-v-west-ham-united-10-april-1999-24023/',
 '/matches/everton-v-coventry-city-11-april-1999-24025/',
 '/matches/leeds-united-v-liverpool-12-april-1999-24027/',
 '/matches/coventry-city-v-middlesbrough-17-april-1999-24029/',
 '/matches/manchester-united-v-sheffield-wednesday-17-april-1999-24031/',
 '/matches/newcastle-united-v-everton-17-april-1999-24032/',
 '/matches/nottingham-forest-v-tottenham-hotspur-17-april-1999-24033/',
 '/matches/liverpool-v-leicester-city-21-april-1999-24039/',
 '/matches/sheffield-wednesday-v-newcastle-united-21-april-1999-24040/',
 '/matches/derby-county-v-southampton-24-april-1999-24043/',
 '/matches/leicester-city-v-coventry-city-24-april-1999-24045/',
 '/matches/tottenham-hotspur-v-west-ham-united-24-april-1999-24047/',
 '/matches/leeds-united-v-manchester-united-25-april-1999-24049/',
 '/matches/newcastle-united-v-middlesbrough-01-may-1999-24056/',
 '/matches/nottingham-forest-v-sheffield-wednesday-01-may-1999-24057/',
 '/matches/southampton-v-leicester-city-01-may-1999-24058/',
 '/matches/west-ham-united-v-leeds-united-01-may-1999-24059/',
 '/matches/liverpool-v-manchester-united-05-may-1999-24063/',
 '/matches/leicester-city-v-newcastle-united-08-may-1999-24069/',
 '/matches/sheffield-wednesday-v-liverpool-08-may-1999-24070/',
 '/matches/middlesbrough-v-manchester-united-09-may-1999-24072/',
 '/matches/coventry-city-v-leeds-united-16-may-1999-24079/',
 '/matches/manchester-united-v-tottenham-hotspur-16-may-1999-24081/',
 '/matches/nottingham-forest-v-leicester-city-16-may-1999-24083/',
 '/matches/southampton-v-everton-16-may-1999-24084/']
