==> Data Cleaning and Fixing/11v11 Data/step2-fix-stadium-data.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("all_match_data_fix1.csv")

stadiums = df['Stadium'].tolist()
stadiums = sorted(set(stadiums))

def isNumber(s):
    for val in [str(i) for i in range(10)]:
        if val in s:
            return True
    return False

nums = 0
for s in stadiums:
    if isNumber(s):
        nums += 1
nums

def replace_stadiums(row):
    stadium = row['Stadium']
    if stadium == "St James' Park Stadium" or stadium == "Sports Direct Arena":
        return "St James' Park"
    elif stadium == "Loftus Road":
        return "Craven Cottage"
    elif stadium == "JJB Stadium":
        return "DW Stadium"
    elif stadium == "City of Manchester Stadium":
        return "Etihad Stadium"
    elif stadium == "Friends Provident St Mary's Stadium":
        return "St Mary's Stadium"
    elif stadium == "Walkers Stadium":
        return "King Power Stadium"
    elif stadium == "KCOM Stadium":
        return "KC Stadium"
    else:
        return stadium

df['Stadium'] = df.apply(replace_stadiums,axis=1)

def select_home_team(match):
    homeTeam = match[0]
    if match[1] == 'v':
        return homeTeam
    elif match[2] == 'v':
        return homeTeam + " " + match[1]
    elif match[3] == 'v':
        return homeTeam + " " + match[1] + " " + match[2]
    elif match[4] == 'v':
        return homeTeam + " " + match[1] + " " + match[2] + " " + match[3]
    else:
        print("Error? : {}".format(match))

team_stadium_dict = dict()
data_links = df['Data Link'].tolist()
homeTeams = [select_home_team(x.split("/")[2].split("-")) for x in data_links]
stadiums = df['Stadium'].tolist()

for i in range(len(df)):
    team = homeTeams[i]
    s = stadiums[i]
    prev_s = team_stadium_dict.get(team)
    if isNumber(s):
        stadiums[i] = prev_s
    else:
        date = data_links[i].split("/")[2].split("-")[-4:-1]
        if s != prev_s and prev_s != None:
            print("{}: Changed from {} -> {}".format(date,prev_s,s))
        team_stadium_dict[team] = s

# Now manually check each stadium change, and ensure all games of the new stadium year are correctly the new stadium
# Firstly need to write out above changes
df['Stadium'] = stadiums
df.to_csv("all_match_data_temp_fix_stadiums.csv",index=False)

# All good
df = pd.read_csv("all_match_data_temp_fix_stadiums.csv")
stadiums = df['Stadium'].tolist()
nums = 0
for s in stadiums:
    if isNumber(s):
        nums += 1
nums

set(stadiums)

def shorten_stadium_names(row):
    s = row['Stadium']
    if s == "Burnden Park, Bolton":
        return "Burnden Park"
    elif s == "Ayresome Park, Middlesbrough":
        return "Ayresome Park"
    elif s == "Bradford and Bingley Stadium":
        return "Valley Parade"
    else:
        return s

df['Stadium'] = df.apply(shorten_stadium_names,axis=1)
df.to_csv("all_match_data_fix2.csv",index=False)
