==> Data Cleaning and Fixing/11v11 Data/step3-split-data-link-col.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("all_match_data_fix2.csv")

data_links = df['Data Link'].tolist()

def month_to_num(month):
    if month == "january":
        return "01"
    elif month == "february":
        return "02"
    elif month == "march":
        return "03"
    elif month == "april":
        return "04"
    elif month == "may":
        return "05"
    elif month == "august":
        return "08"
    elif month == "september":
        return "09"
    elif month == "october":
        return "10"
    elif month == "november":
        return "11"
    elif month == "december":
        return "12"
    else:
        print("Invalid Month: {}".format(month))

def convert_date_format(date_wrong):
    day_num = date_wrong[0]
    month = date_wrong[1]
    month_num = month_to_num(month)
    year_num = date_wrong[2]
    date = day_num + month_num + year_num
    return date

def split_data_link(data_link):
    useful_info = data_link.split("/")[2].split("-")
    eleven_id = useful_info[-1]
    date_wrong_format = useful_info[-4:-1]
    date = convert_date_format(date_wrong_format)
    teams_data = data_link.split("/")[2].split(date_wrong_format[0])[0]
    teams_data_split = teams_data.split("-v-")
    home_team = teams_data_split[0].replace("-"," ")
    away_team = teams_data_split[1].replace("-"," ")
    away_team = away_team.strip()
    return (eleven_id,date,home_team,away_team)

print(split_data_link(data_links[0]))

def split_all_data_links(data_links):
    all_ids = []
    all_dates = []
    all_home = []
    all_away = []
    for data_link in data_links:
        (eleven_id,date,home_team,away_team) = split_data_link(data_link)
        all_ids.append(eleven_id)
        all_dates.append(date)
        all_home.append(home_team)
        all_away.append(away_team)
    info_dict = {'11v11 ID':all_ids,'Date':all_dates,'Home Team':all_home,'Away Team':all_away}
    return info_dict

info_dict = split_all_data_links(data_links)
df_extra = pd.DataFrame(info_dict)
df_extra.head()

df_final = pd.concat([df,df_extra],axis=1)
df_final.head()

df_final.drop("Data Link",inplace=True,axis=1)
df_final.head()

df_final = df_final[['11v11 ID','Date','Home Team','Away Team','Stadium','Referee','HomePlayers','AwayPlayers','HomeScorers','AwayScorers']]
df_final.head()

df_final.columns = ['11v11 ID','Date','Home Team','Away Team','Stadium','Referee','Home Players','Away Players','Home Scorers','Away Scorers']
df_final.head()

df_final.to_csv("all_match_data_fix3.csv",index=False)
