==> Data Cleaning and Fixing/11v11 Data/step4-code-lower-case-names.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("all_match_data_fix3.csv")

def lower_case(row):
    s = row['Stadium']
    return s.lower()

df['Stadium'] = df.apply(lower_case,axis=1)

def lower_case2(row):
    r = row['Referee']
    return r.lower()

df['Referee'] = df.apply(lower_case2,axis=1)

df.head()

df.to_csv("all_match_data_fix4.csv",index=False)
