==> Data Cleaning and Fixing/11v11 Data/step5-reduce-team-names.ipynb <==
import pandas as pd

df = pd.read_csv("4-data_all_match_data.csv")

len(set(df['Home Team'].tolist()))

def replace_names(row,col):
    team = row[col]
    if team == "afc bournemouth":
        return "bournemouth"
    elif team == "birmingham city":
        return "birmingham"
    elif team == "blackburn rovers":
        return "blackburn"
    elif team == "bolton wanderers":
        return "bolton"
    elif team == "bradford city":
        return "bradford"
    elif team == "brighton and hove albion":
        return "brighton"
    elif team == "cardiff city":
        return "cardiff"
    elif team == "charlton athletic":
        return "charlton"
    elif team == "coventry city":
        return "coventry"
    elif team == "derby county":
        return "derby"
    elif team == "huddersfield town":
        return "huddersfield"
    elif team == "hull city":
        return "hull"
    elif team == "ipswich town":
        return "ipswich"
    elif team == "leeds united":
        return "leeds"
    elif team == "leicester city":
        return "leicester"
    elif team == "manchester city":
        return "man city"
    elif team == "manchester united":
        return "man utd"
    elif team == "newcastle united":
        return "newcastle"
    elif team == "norwich city":
        return "norwich"
    elif team == "oldham athletic":
        return "oldham"
    elif team == "queens park rangers":
        return "qpr"
    elif team == "sheffield united":
        return "sheffield utd"
    elif team == "sheffield wednesday":
        return "sheffield wed"
    elif team == "stoke city":
        return "stoke"
    elif team == "swansea city":
        return "swansea"
    elif team == "swindon town":
        return "swindon"
    elif team == "tottenham hotspur":
        return "spurs"
    elif team == "west bromwich albion":
        return "west brom"
    elif team == "west ham united":
        return "west ham"
    elif team == "wigan athletic":
        return "wigan"
    elif team == "wolverhampton wanderers":
        return "wolves"
    else:
        return team

df['Home Team'] = df.apply(replace_names,args=('Home Team',),axis=1)
df['Away Team'] = df.apply(replace_names,args=('Away Team',),axis=1)
set(df['Home Team'].tolist())

df.head()

df.to_csv("all_match_data_fix5.csv",index=False)
