==> Data Cleaning and Fixing/11v11 Data/step6-analyse-progress-so-far.ipynb <==
import pandas as pd
df = pd.read_csv("5-data_all_match_data.csv")

set(df['Stadium'])

list(df.columns.values)

set(df['Referee'])

df.iloc[1]['Home Players']

print(len(df))

all_teams = set(df['Home Team'].tolist())
all_teams_dict = {}
for team in all_teams:
    all_teams_dict[team] = 0
for i in range(len(df)):
    row = df.iloc[i]
    hteam = row['Home Team']
    ateam = row['Away Team']
    all_teams_dict[hteam] += 1
    all_teams_dict[ateam] += 1
print(all_teams_dict)
